refactor: rename ow-core to core
