+++ packages/core/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@dotkomonline/ow-core",
+  "name": "@dotkomonline/core",
   "version": "1.0.0",
   "main": "./src/index.ts",
   "types": "./src/index.ts",
