+++ packages/core/src/errors/errors.ts
