+++ packages/core/src/modules/event/attendance-repository.ts
