+++ packages/core/src/modules/event/attendance-service.ts
