+++ packages/ow-gateway-trpc/src/modules/company/company-router.ts
@@ -1,7 +1,7 @@
 import { z } from "zod"
 import { t } from "../../trpc"
 import { CompanyWriteSchema } from "@dotkomonline/types"
-import { PaginateInputSchema } from "@dotkomonline/ow-core"
+import { PaginateInputSchema } from "@dotkomonline/core"
 
 export const companyRouter = t.router({
   create: t.procedure.input(CompanyWriteSchema).mutation(({ input, ctx }) => {
